server/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
numpy==2.0.2
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
spotipy==2.24.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
